requirements.txt
fastapi
uvicorn
pyspark
psycopg2
kafka-python
scikit-learn
numpy
pandas
pydantic
